Makefile: statically linked musl libc for Kobo
Enables usage of the musl libc for the KOBO target, for usage with a
toolchaing for target "armv7-kobo-linux-musleabihf" (which can be built
with crosstool-ng).
All libraries, including libgcc and musl libc, are now statically
linked, using "-static" as a linker flag.
Remove glibc as a thirdparty library for the KOBO target, and the
actions for including any dynamic libraries KoboRoot.tgz.