Screen/EGL/TopCanvas: probe for usable DRI device
Instead of trying to initialise only the first DRI card device, try to
initialise DRM / GBM context on any card until it succeeds.

This is useful, because on many devices (probably even on most), there
are multiple /dev/dri/card* device files, and the usable device is not
necessarily card0.

It is still possible to specify a particular DRI device using the
environment variable "DRI_DEVICE".