Screen/EGL/TopCanvas: use ioctl() instead of libdrm functions
Instead of using the API of libdrm, interact directly with the Linux
kernel using ioctl calls for interacting with the Direct Rendering
Infrastructure. This avoids some overhead, and we do not have to link
against libdrm.

However, libdrm ist still needed at compile time, because we still need
the Linux kernel DRM data type declarations from the libdrm headers.
This could normally be done using the UAPI kernel headers, but these are
not distributed in all Linux distributions.