build/python/libs: libpng: use CMake instead of Autotools
When compiling with Autotools, this library uses the "libtool" script,
which strips the "-static" linker flag (like CURL and libjpeg-turbo).
There is no easy workaround for this, but switching to libpng's CMake
build system solves this problem.

Furthermore, enabling ARM NEON conditionally is not required, because
the CMake PNG_ARM_NEON option is ignored anyway for non-ARM targets.